+++ packages/node/src/index.ts
@@ -1,21 +1,20 @@
 /**
- * This is the API Reference for the Experiment JS Server SDK.
- * For more details on implementing this SDK, view the documentation
- * [here](https://amplitude-lab.readme.io/docs/javascript-server-sdk).
+ * This is the API Reference for the Experiment Node.js Server SDK.
+ * For more details on implementing this SDK, view the [documentation]
+ * (https://docs.developers.amplitude.com/experiment/sdks/nodejs-sdk).
  * @module experiment-node-server
  */
 
-export { AmplitudeCookie } from './amplitude';
-export { ExperimentClient } from './client';
-export { ExperimentConfig, Defaults } from './config';
+export { AmplitudeCookie } from './cookie';
+export { ExperimentClient } from './remote/client';
+export { ExperimentConfig, Defaults } from './types/config';
 export { Experiment } from './factory';
 export { ExperimentUser } from './types/user';
 export { Variant, Variants } from './types/variant';
 
 export { LocalEvaluationClient } from './local/client';
-export { LocalEvaluationConfig } from './config';
+export { LocalEvaluationConfig } from './types/config';
 export { FlagConfigFetcher } from './local/fetcher';
 export { FlagConfigPoller } from './local/poller';
-export { FlagConfigEvaluator } from './local/evaluator';
 export { InMemoryFlagConfigCache } from './local/cache';
 export { FlagConfig, FlagConfigCache } from './types/flag';
