+++ packages/node/src/local/fetcher.ts
@@ -1,8 +1,7 @@
-import { LocalEvaluationDefaults } from '../config';
-import { ConsoleLogger } from '../logger/console';
-import { FetchHttpClient } from '../transport/http';
+import { LocalEvaluationDefaults } from '../types/config';
 import { FlagConfig } from '../types/flag';
 import { HttpClient } from '../types/transport';
+import { ConsoleLogger } from '../util/logger';
 import { Logger } from '../util/logger';
 
 const FLAG_CONFIG_TIMEOUT = 5000;
