--- packages/node/src/logger/console.ts
@@ -1,17 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { Logger } from '../util/logger';
-
-export class ConsoleLogger implements Logger {
-  private readonly isDebug: boolean;
-  constructor(isDebug: boolean) {
-    this.isDebug = isDebug;
-  }
-  debug(message?: string, ...optionalParams: any[]): void {
-    if (this.isDebug) {
-      console.debug(message, ...optionalParams);
-    }
-  }
-  error(message?: string, ...optionalParams: any[]): void {
-    console.error(message, ...optionalParams);
-  }
-}
